+++ hydromt_wflow/wflow.py
@@ -19,6 +19,7 @@ import numpy as np
 import pandas as pd
 import pyflwdir
 import pyproj
+import shapely
 import toml
 import xarray as xr
 from dask.diagnostics import ProgressBar
@@ -224,6 +225,21 @@ class WflowModel(GridModel):
         # retrieve global data (lazy!)
         ds_org = self.data_catalog.get_rasterdataset(hydrography_fn)
 
+        # Check on resolution (degree vs meter) depending on ds_org res/crs
+        scale_ratio = int(np.round(res / ds_org.raster.res[0]))
+        if scale_ratio < 1:
+            raise ValueError(
+                f"The model resolution {res} should be \
+larger than the {hydrography_fn} resolution {ds_org.raster.res[0]}"
+            )
+        if ds_org.raster.crs.is_geographic:
+            if res > 1:  # 111 km
+                raise ValueError(
+                    f"The model resolution {res} should be smaller than 1 degree \
+(111km) for geographic coordinate systems. "
+                    "Make sure you provided res in degree rather than in meters."
+                )
+
         # get basin geometry and clip data
         kind, region = hydromt.workflows.parse_region(region, logger=self.logger)
         xy = None
@@ -249,25 +265,15 @@ class WflowModel(GridModel):
         if geom is not None and geom.crs is None:
             raise ValueError("wflow region geometry has no CRS")
 
+        # Set the basins geometry
         ds_org = ds_org.raster.clip_geom(geom, align=res, buffer=10)
         ds_org.coords["mask"] = ds_org.raster.geometry_mask(geom)
         self.logger.debug("Adding basins vector to geoms.")
-        self.set_geoms(geom, name="basins")
 
-        # Check on resolution (degree vs meter) depending on ds_org res/crs
-        scale_ratio = int(np.round(res / ds_org.raster.res[0]))
-        if scale_ratio < 1:
-            raise ValueError(
-                f"The model resolution {res} should be \
-larger than the {hydrography_fn} resolution {ds_org.raster.res[0]}"
-            )
-        if ds_org.raster.crs.is_geographic:
-            if res > 1:  # 111 km
-                raise ValueError(
-                    f"The model resolution {res} should be smaller than 1 degree \
-(111km) for geographic coordinate systems. "
-                    "Make sure you provided res in degree rather than in meters."
-                )
+        # Set name based on scale_factor
+        if scale_ratio != 1:
+            self.set_geoms(geom, name="basins_highres")
+
         # setup hydrography maps and set staticmap attribute with renamed maps
         ds_base, _ = workflows.hydrography(
             ds=ds_org,
@@ -295,6 +301,8 @@ larger than the {hydrography_fn} resolution {ds_org.raster.res[0]}"
         # Rename and add to grid
         rmdict = {k: v for k, v in self._MAPS.items() if k in ds_base.data_vars}
         self.set_grid(ds_base.rename(rmdict))
+        # Call basins once to set it
+        self.basins
 
         # setup topography maps
         ds_topo = workflows.topography(
@@ -427,6 +435,14 @@ larger than the {hydrography_fn} resolution {ds_org.raster.res[0]}"
         """
         self.logger.info("Preparing river maps.")
 
+        # Check that river_upa threshold is bigger than the maximum uparea in the grid
+        if river_upa > float(self.grid[self._MAPS["uparea"]].max()):
+            raise ValueError(
+                f"river_upa threshold {river_upa} should be larger than the maximum \
+uparea in the grid {float(self.grid[self._MAPS['uparea']].max())} in order to create \
+river cells."
+            )
+
         rivdph_methods = ["gvf", "manning", "powlaw"]
         if rivdph_method not in rivdph_methods:
             raise ValueError(f'"{rivdph_method}" unknown. Select from {rivdph_methods}')
@@ -471,11 +487,12 @@ Select from {routing_options}.'
         df = self.data_catalog.get_dataframe(rivman_mapping_fn)
         # max streamorder value above which values get the same N_River value
         max_str = df.index[-2]
+        nodata = df.index[-1]
         # if streamorder value larger than max_str, assign last value
         strord = strord.where(strord <= max_str, max_str)
         # handle missing value (last row of csv is mapping of missing values)
-        strord = strord.where(strord != strord.raster.nodata, -999)
-        strord.raster.set_nodata(-999)
+        strord = strord.where(strord != strord.raster.nodata, nodata)
+        strord.raster.set_nodata(nodata)
         ds_nriver = workflows.landuse(
             da=strord,
             ds_like=self.grid,
@@ -913,8 +930,7 @@ to run setup_river method first.'
             fn_map = f"{lulc_fn}_mapping_default"
         else:
             fn_map = lulc_mapping_fn
-        if not isfile(fn_map) and fn_map not in self.data_catalog:
-            raise ValueError(f"LULC mapping file not found: {fn_map}")
+
         # read landuse map to DataArray
         da = self.data_catalog.get_rasterdataset(
             lulc_fn, geom=self.region, buffer=2, variables=["landuse"]
@@ -1797,7 +1813,7 @@ Using default storage/outflow function parameters."
             kwargs.update(predicate="contains")
         gdf_org = self.data_catalog.get_geodataframe(
             waterbodies_fn,
-            geom=self.basins,
+            geom=self.basins_highres,
             handle_nodata=NoDataStrategy.IGNORE,
             **kwargs,
         )
@@ -2052,7 +2068,7 @@ added to glacierstore [-]
         self.logger.info("Preparing glacier maps.")
         gdf_org = self.data_catalog.get_geodataframe(
             glaciers_fn,
-            geom=self.basins,
+            geom=self.basins_highres,
             predicate="intersects",
             handle_nodata=NoDataStrategy.IGNORE,
         )
@@ -2252,6 +2268,7 @@ one variable and variables list is not provided."
             time_tuple=(starttime, endtime),
             variables=["precip"],
         )
+        precip = precip.astype("float32")
 
         if chunksize is not None:
             precip = precip.chunk({"time": chunksize})
@@ -2264,6 +2281,7 @@ one variable and variables list is not provided."
                 buffer=2,
                 variables=["precip"],
             )
+            clim = clim.astype("float32")
 
         precip_out = hydromt.workflows.forcing.precip(
             precip=precip,
@@ -2397,7 +2415,7 @@ either {'temp' [°C], 'temp_min' [°C], 'temp_max' [°C], 'wind' [m/s], 'rh' [%]
             else:
                 methods = [
                     "debruin",
-                    "makking",
+                    "makkink",
                     "penman-monteith_rh_simple",
                     "penman-monteith_tdew",
                 ]
@@ -2415,6 +2433,8 @@ either {'temp' [°C], 'temp_min' [°C], 'temp_max' [°C], 'wind' [m/s], 'rh' [%]
         )
         if chunksize is not None:
             ds = ds.chunk({"time": chunksize})
+        for var in ds.data_vars:
+            ds[var] = ds[var].astype("float32")
 
         dem_forcing = None
         if dem_forcing_fn is not None:
@@ -2424,6 +2444,7 @@ either {'temp' [°C], 'temp_min' [°C], 'temp_max' [°C], 'wind' [m/s], 'rh' [%]
                 buffer=2,
                 variables=["elevtn"],
             ).squeeze()
+            dem_forcing = dem_forcing.astype("float32")
 
         temp_in = hydromt.workflows.forcing.temp(
             ds["temp"],
@@ -2545,6 +2566,7 @@ either {'temp' [°C], 'temp_min' [°C], 'temp_max' [°C], 'wind' [m/s], 'rh' [%]
             variables=["pet"],
             time_tuple=(starttime, endtime),
         )
+        pet = pet.astype("float32")
 
         pet_out = workflows.forcing.pet(
             pet=pet,
@@ -3647,6 +3669,7 @@ Run setup_soilmaps first"
     def write_geoms(
         self,
         geom_fn: str = "staticgeoms",
+        precision: int | None = None,
     ):
         """Write geoms in <root/geom_fn> in GeoJSON format."""
         # to write use self.geoms[var].to_file()
@@ -3654,7 +3677,21 @@ Run setup_soilmaps first"
             raise IOError("Model opened in read-only mode")
         if self.geoms:
             self.logger.info("Writing model staticgeom to file.")
+            # Set projection to 1 decimal if projected crs
+            _precision = precision
+            if precision is None:
+                if self.crs.is_projected:
+                    _precision = 1
+                else:
+                    _precision = 6
+            grid_size = 10 ** (-_precision)
             for name, gdf in self.geoms.items():
+                # TODO change to geopandas functionality once geopandas 1.0.0 comes
+                # See https://github.com/geopandas/geopandas/releases/tag/v1.0.0-alpha1
+                gdf.geometry = shapely.set_precision(
+                    gdf.geometry,
+                    grid_size=grid_size,
+                )
                 fn_out = join(self.root, geom_fn, f"{name}.geojson")
                 gdf.to_file(fn_out, driver="GeoJSON")
 
@@ -4089,7 +4126,7 @@ change name input.path_forcing "
         if len(fns) > 0:
             for fn in fns:
                 name = basename(fn).split(".")[0]
-                tbl = pd.read_csv(fn)
+                tbl = pd.read_csv(fn, float_precision="round_trip")
                 self.set_tables(tbl, name=name)
 
     def write_tables(self):
@@ -4168,11 +4205,19 @@ change name input.path_forcing "
                 .set_index("value")
                 .sort_index()
             )
-            gdf.index.name = self._MAPS["basins"]
             self.set_geoms(gdf, name="basins")
         else:
             self.logger.warning(f"Basin map {self._MAPS['basins']} not found in grid.")
             gdf = None
+        return gdf
+
+    @property
+    def basins_highres(self):
+        """Returns a high resolution basin(s) geometry."""
+        if "basins_highres" in self.geoms:
+            gdf = self.geoms["basins_highres"]
+        else:
+            gdf = self.basins
         return gdf
 
     @property